Refactor Excel validators to be part of UnisContent object
