+++ src/test/java/no/sikt/nva/scrapers/ExcelScraperTest.java
@@ -1,7 +1,7 @@
 package no.sikt.nva.scrapers;
 
-import static no.sikt.nva.validators.ExcelRowValidator.ERROR_MESSAGE_INVALID_EMBARGO_FORMAT;
-import static no.sikt.nva.validators.ExcelScraperValidator.ERROR_MESSAGE_EMPTY_SHEET;
+import static no.sikt.nva.UnisContent.ERROR_MESSAGE_INVALID_EMBARGO_FORMAT;
+import static no.sikt.nva.UnisContentScrapingResult.ERROR_MESSAGE_EMPTY_SHEET;
 import static no.unit.nva.testutils.RandomDataGenerator.randomString;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
